+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/tests/Helpers/StringConversionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace GraphQLAPI\GraphQLAPI\Misc;
+namespace GraphQLAPI\GraphQLAPI\Helpers;
 
 use GraphQLAPI\GraphQLAPI\Services\Helpers\StringConversion;
 use PoP\Engine\AbstractTestCase;
